move with_credentials method to PriceloopClient object
